Locurilla goty

+++ BackEnd/src/main/java/com/example/crudrapido/Controlador/IngredienteControlador.java
@@ -1,11 +1,19 @@
 package com.example.crudrapido.Controlador;
 import java.util.List;
 
-import com.example.crudrapido.Repository.IngredienteRepositorio;
-import com.example.crudrapido.model.Ingrediente;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.example.crudrapido.Repository.IngredienteRepositorio;
+import com.example.crudrapido.Repository.ProductoRepositorio;
+import com.example.crudrapido.model.Ingrediente;
+import com.example.crudrapido.model.Producto;
 
 
 @RestController
@@ -15,6 +23,8 @@ public class IngredienteControlador {
 
     @Autowired
     private IngredienteRepositorio repositorio;
+    @Autowired
+    private ProductoRepositorio productoRepositorio;
 
     @GetMapping("/ingredientes")
     public List<Ingrediente> ListarIngredientes(){
@@ -39,8 +49,23 @@ public class IngredienteControlador {
             {return null;}
     }
 
-    @PostMapping("/ingredientes/{id}")
+@PostMapping("/ingredientes/{id}")
   public void eliminarIngrediente(@PathVariable Integer id){
-      repositorio.deleteById(id);
+      List<Producto> x =productoRepositorio.findAll();
+      boolean verificar=true;
+      for (Producto p : x) {
+        if(p.getIngrediente1().equals(id)||p.getIngrediente2().equals(id)||p.getIngrediente3().equals(id)||p.getIngrediente4().equals(id)||p.getIngrediente5().equals(id)){
+        System.out.println("llegué al if xd");
+        verificar=false;
+        }
+        }
+        if(verificar)
+        {
+          System.out.println("llegué al principio del else");
+          repositorio.deleteById(id);
+          System.out.println("llegué al final else");
+        }
+      
     }
 }
+
